requirements.txt
amqp==1.4.9
anyjson==0.3.3
args==0.1.0
beautifulsoup4==4.4.1
billiard==3.3.0.22
boto==2.39.0
celery==3.1.20
clint==0.5.1
csscompressor==0.9.4
dj-database-url==0.4.0
Django==1.9.2
django-appconf==1.0.1
django-compress==1.0.1
django-compressor==2.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-extras==0.3
django-localflavor==1.2
django-middleware-extras==0.1.2
django-pdb==0.4.2
django-pipeline==1.6.6
django-sass-processor==0.3.4
django-storages==1.4
eventlet==0.18.4
extras==0.0.3
greenlet==0.4.9
gunicorn==19.4.5
inflection==0.3.1
jsmin==2.2.0
jsonfield==1.0.3
kombu==3.0.33
libsass==0.10.1
psycopg2==2.6.1
python-postmark==0.4.7
python-postmark-inbound==1.1.0
pytz==2015.7
PyYAML==3.11
rcssmin==1.0.6
redis==2.10.5
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
sqlparse==0.1.18
sunlight==1.2.9
